# WIP: speed up JumpProblem conversion 2