https://gist.github.com/isaacsas/593f9ce7f40adf05b6969ace8fe8a831

with

```julia
const nedgess = 30
const nvertss = 1000
```
gives

```julia
  0.000425 seconds (14.00 k allocations: 641.219 KiB)
Completed: ReactionSystem
  0.033972 seconds (519.34 k allocations: 16.500 MiB)
Completed: convert(JumpSystem)
  0.025654 seconds (585.68 k allocations: 15.495 MiB)
Completed: DiscreteProblem
  2.788977 seconds (2.35 M allocations: 2.143 GiB, 5.16% gc time)
Completed: JumpProblem
  0.000594 seconds (30 allocations: 353.297 KiB)
```

This is probably really bad. I'm thinking of better hash-table operations for the purposes here, hopefully we can turn something from `O(n^2)` to `O(n)`