@shashi 2020-11-01
This is pretty interesting, most of the time is spent in substitute (the call to `subber` in `assemble_maj`), and not the dictionary construction. I guess this should be even faster than what we had before though.