@isaacsas 2020-11-01
It was pretty bad before, though in the other open PR I had gotten it down a fair amount for everything but the final JumpProblem construction, which was performing badly due to the dependency graph construction among other things.