@ChrisRackauckas 2020-11-01
3 seconds still doesn't seem great.