@shashi 2020-11-01
Got another 10x

it was doing wasteful dict construction.

```
  0.000401 seconds (14.00 k allocations: 641.219 KiB)
Completed: ReactionSystem
  0.030159 seconds (519.34 k allocations: 16.500 MiB)
Completed: convert(JumpSystem)
  0.024397 seconds (585.68 k allocations: 15.495 MiB)
Completed: DiscreteProblem
  0.311764 seconds (2.18 M allocations: 57.016 MiB)
Completed: JumpProblem
  0.000451 seconds (30 allocations: 353.297 KiB)
Completed: solve
```