@shashi 2020-11-01
I think we should only leave `substitue(expr, dict::Dict)` method since everything essentially calls that. This way it's much easier to reason about substitute's performance.