@shashi 2020-11-01
Okay another 3x

```
  0.093674 seconds (2.18 M allocations: 57.018 MiB)
```