@shashi 2020-11-01
Now with
```julia
const nedgess = 3000
const nvertss = 100000
```
```julia
  0.058344 seconds (1.50 M allocations: 64.072 MiB)
Completed: ReactionSystem
  5.699431 seconds (44.17 M allocations: 1.322 GiB, 44.66% gc time)
Completed: convert(JumpSystem)
  4.851998 seconds (52.66 M allocations: 1.283 GiB, 18.10% gc time)
Completed: DiscreteProblem
 17.310221 seconds (211.11 M allocations: 5.235 GiB, 12.65% gc time)
Completed: JumpProblem
  0.055617 seconds (30 allocations: 34.334 MiB)
Completed: solve
retcode: Default
Interpolation: Piecewise constant interpolation
t: 6-element Array{Float64,1}:
  0.0
  0.1
  0.2
  0.3
  0.4
 20.0
u: 6-element Array{Array{Int64,1},1}:
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
 [0, 1, 1, 1, 1, 1, 1, 1, 1, 1  …  0, 0, 0, 0, 0, 0, 0, 0, 0, 0]

```