@shashi 2020-11-01
```
function states(sys::AbstractSystem)
    unique(isempty(sys.systems) ?
           setdiff(sys.states, value.(sys.pins)) :
           [sys.states;reduce(vcat,namespace_variables.(sys.systems))])
end
```

This is pretty expensive, especially `setdiff`. We should cache this.

Similarly `get_variables!` does `unique` which is expensive too.