@isaacsas 2020-11-01
Could also just return `states` if `pins` and `systems` are both empty.