@shashi 2020-11-02
> Could also just return states if pins and systems are both empty. I think that’s what I had put in in the other PR.

That's kind of a kludge but ok!

Do you know what's causing the test to fail now?