@isaacsas 2020-11-03
@shashi Before we merge, can copy the latest timings over here so we have a reference for future comparisons?